+++ ReleaseNotes.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## Version 1.0beta1
+* support NDI sources on the same PC
+* change advertised Camera names to "@CAM..."
+* filter ndi selector sources bases on camera name
+* fix formatting with long NDI source names
+
 ## Version 0.85
 * Rearrange the Set Camera/Set PTZ UI to improve intuition
 * Fix a couple exceptions